# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ayna API
  version: 1.0.0
  description: |
    The consumer REST surface for Ayna.

    This file is the contract between the Flutter client and the Go backend.
    It is not documentation written after the fact — it is the source both
    sides generate from, and it is checked against the Go domain enums by
    `internal/contracts` so the two cannot drift apart silently.

    ## Three things that shape almost every response here

    **Severity has five values, not four.** `unknown` is not padding: a concern
    the analysis could not assess must be representable. Rendering an
    unassessed concern as `none` would tell someone their skin is clear when
    nothing ever looked at it, which is the worst thing this API can say.

    **Free-tier reports have no `skinAge` and no `heatmap`.** Those come from
    the specialist vendor, and the free tier runs a cheaper model (PD-3). Both
    fields are nullable, and clients must render the absence rather than
    treating it as an error or substituting a placeholder.

    **Nothing here is a diagnosis.** Every response carrying a result also
    carries `disclaimer`, and clients are required to display it (NFR-6). It
    is in the payload rather than hardcoded in the app so the wording can be
    corrected after review without shipping a release.

  contact:
    name: Ayna
  license:
    name: Proprietary
servers:
  - url: https://api.ayna.app/v1
    description: Production
...
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /users:
    post:
      tags: [Identity]
      operationId: createUser
      summary: Create the local profile after Auth0 sign-up
      description: |
        Called once, after Auth0 has authenticated the user for the first time.
        Auth0 answers "who is this person"; this endpoint decides whether they
        may have an account at all.

        **The age gate lives here (PD-1).** A caller under 18 is refused and no
        user is created — refusing at creation means there is never an underage
        record to purge later. `birthYear` alone is collected, never a full date
        of birth: the year satisfies both the gate and the skin-age comparison,
        and a full date is more identifying for no added benefit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [birthYear, timezone]
              properties:
                birthYear:
                  type: integer
                  minimum: 1900
                  maximum: 2100
                  example: 1999
                timezone:
                  type: string
                  description: |
                    IANA zone. Not cosmetic — it defines the week boundaries the
                    streak is computed in (PD-5). UTC boundaries would credit a
                    2am scan in Karachi to the previous week, which is invisible
                    to the user and reads as a bug.
                  example: Asia/Karachi
                displayName:
                  type: string
                  maxLength: 60
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: |
            Under the minimum age. Deliberately distinct from a validation
            error: the client shows a specific, non-punitive screen for this,
            not a generic failure.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
              example:
                type: https://api.ayna.app/problems/under-minimum-age
                title: Below the minimum age
                status: 403
                detail: Ayna is available to people aged 18 and over.
        "409":
          description: A profile already exists for this identity
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /users/me:
    get:
      tags: [Identity]
      operationId: getCurrentUser
      summary: The signed-in user, with entitlement
      description: |
        The client's source of truth for what to show. `entitlement` decides
        whether the report screen expects a heatmap and whether the scan button
        is enabled, so this is fetched on launch and after any purchase.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404":
          description: Authenticated but no local profile yet — call POST /users
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

    patch:
      tags: [Identity]
      operationId: updateCurrentUser
      summary: Change the photo retention policy (NFR-4)
      description: |
        PATCH rather than PUT: this replaces one field of the profile, and a
        PUT would imply the client may replace `birthYear`, which PD-1 requires
        to be immutable after the age gate.

        `photoRetentionDays` is required in the body and `null` is a legal
        value meaning "keep indefinitely". There is no partial-update
        ambiguity to resolve because the resource has exactly one mutable
        field.

        Setting a shorter policy does not delete anything synchronously. The
        retention sweep runs on its own schedule, so images already past the
        new window survive until it next runs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [photoRetentionDays]
              properties:
                photoRetentionDays:
                  type: integer
                  nullable: true
                  enum: [7, 30, 365, null]
                  description: Days to keep photos, or null to keep them indefinitely.
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Not one of the offered retention options
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Authenticated but no local profile yet
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

    delete:
      tags: [Identity]
      operationId: deleteCurrentUser
      summary: Delete the account, photos, and derived data (FR-11)
      description: |
        Returns 202, not 204: deletion is asynchronous because it spans
        Postgres, object storage and the vendor's side. The response carries
        the SLA so the client can state a real timeframe rather than "soon".
      responses:
        "202":
          description: Deletion accepted
          content:
            application/json:
              schema:
                type: object
                required: [requestedAt, completesBy]
                properties:
                  requestedAt: { type: string, format: date-time }
                  completesBy:
                    type: string
                    format: date-time
                    description: Outer bound of the purge SLA.
  /consent:
    post:
      tags: [Identity]
      operationId: grantConsent
      summary: Record explicit consent to photo processing (FR-1)
      description: |
        A scan is refused until this exists. Consent is **versioned** — when the
        wording changes, prior consent does not silently cover the new terms,
        and the user is asked again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policyVersion, granted]
              properties:
                policyVersion:
                  type: string
                  example: "2026-08-01"
                granted:
                  type: boolean
                  description: |
                    Revocation posts `false` rather than deleting the record.
                    An audit trail that can be erased is not an audit trail.
      responses:
        "201":
          description: Consent recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsentRecord" }

  # ---------------------------------------------------------------------------
  # Scans — the riskiest flow, specced in full
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0-issued access token.
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details. Every error in this API uses this shape.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: |
            Safe to show a user. Errors say what went wrong and what to do
            about it — never an apology, never a stack trace.
        correlationId:
          type: string
          description: Ties this response to the server logs (NFR-9).
...
    Entitlement:
      type: object
      required: [tier, scansRemaining, periodResetsAt]
      properties:
        tier: { type: string, enum: [free, plus, pro] }
        scansRemaining:
          type: integer
          minimum: 0
          description: Remaining in the current period.
        periodResetsAt: { type: string, format: date-time }
        hasHeatmap:
          type: boolean
          description: |
            Stated explicitly rather than inferred from `tier`. The client
            should not encode which tiers include which features — that mapping
            changes with pricing, and a stale app would then render an empty
            heatmap slot or hide one the user paid for.
        hasSkinAge: { type: boolean }
    User:
      type: object
      required: [id, birthYear, timezone, entitlement, hasConsent]
      properties:
        id: { type: string, example: usr_01J8XK4M2P9QRSTVWXYZABCDEF }
        displayName: { type: string, nullable: true }
        birthYear: { type: integer }
        timezone: { type: string, example: Asia/Karachi }
        hasConsent:
          type: boolean
          description: Whether a current, un-revoked consent record exists for the live policy version.
        entitlement: { $ref: "#/components/schemas/Entitlement" }
        streakWeeks:
          type: integer
          description: Consecutive ISO weeks with at least one completed scan (PD-5).
        photoRetentionDays:
          type: integer
          nullable: true
          enum: [7, 30, 365, null]
          description: |
            How long photos are kept before the retention sweep clears them
            (NFR-4). `null` means keep indefinitely, and is a deliberate user
            choice rather than an unset field: new accounts default to 30.

            Only the image is removed. Scores and the trend survive, because a
            user asking for their photos to be deleted is not asking for their
            history to be erased.
    ConsentRecord:
      type: object
      required: [policyVersion, granted, recordedAt]
      properties:
        policyVersion: { type: string }
        granted: { type: boolean }
        recordedAt: { type: string, format: date-time }
...